+++ resources/from2_1.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 
 ##
-# Acuparse 2.1 Migration Script
+# Acuparse Installation Script
+# 2.1 Apache Migration Tasks
 # @copyright Copyright (C) 2015-2018 Maxwell Power
 # @author Maxwell Power <max@acuparse.com>
 # @license MIT