+++ resources/le.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 
 ##
-# Acuparse Let's Encrypt Installer
+# Acuparse Installation Script
+# Let's Encrypt Installer
 # @copyright Copyright (C) 2015-2018 Maxwell Power
 # @author Maxwell Power <max@acuparse.com>
 # @license MIT