+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Acuparse Installer
+# Acuparse Installation Script
 
 # Change Log
 All notable changes to this project will be documented in this file.
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.2.0] - 2018-07-06
+### Changed
+- Removed Ubuntu 16.04 LTS support and replaced with 18.04 LTS.
+- Support for PHP 7.2.
+
 ## [1.1.0] - 2018-01-31
 ### Added
-- SSL & Let's Encrypt Support
+- SSL & Let's Encrypt Support.
 
 ## [1.0.1] - 2017-08-08
 ### Changed
@@ -18,4 +23,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.0] - 2017-03-13
 ### Added
-- Initial open source release
+- Initial open source release.